Refactor Fluid api common part
